+++ Draki.Tests/Pages/Colors.cs
@@ -3,7 +3,7 @@
     public static class Colors
     {
         public const string ALICE_BLUE = "rgb(240, 248, 255)";
-        public const string RED = "rgb(255, 248, 255)";
+        public const string RED = "rgb(255, 0, 0)";
         public const string SILVER = "rgb(230, 230, 230)";
         public const string DARK_BLUE = "rgb(40, 96, 144)";
     }